fix(sync): 谷歌云盘登录走全应用代理出口，授权超时不再误判为登录过期 (BUG-1348)
用户报「谷歌网页版显示登入，返回软件显示超时」并问是不是 API 限额。不是：
日志 10 次 signIn 全败，8 次是 token 交换直连 oauth2.googleapis.com 被切
(errno=121)，2 次是 loopback 回调没回来。登录阶段还没调过任何 Drive API。

四条独立根因：

1. 桌面登录横跨两个网络栈，第二段是硬直连。浏览器那半程走系统代理（所以网页
   显示已登入），app 拿授权码换 token 用裸 http.Client()——Dart 的 HttpClient
   默认 findProxy 为空，连 HTTPS_PROXY 都不读，更不读注册表里的系统代理。仓库
   早有正确实现 applyUpdateProxy，但它住在 update_checker 的 part 文件里，而
   结构守卫钉死「part 不得自带 import」，同步层根本没法复用。
   → 代理层提取为独立库 src/utils/net/app_proxy.dart（applyAppProxy），成为全
     应用唯一出口真相源；新增进程级 appUserProxyReader 由 AppModel 接到偏好，
     不再沿调用链穿参（穿漏一处就是一条不走代理的暗路）。
2. 同步层出站请求没有连接超时，吊到 OS 默认。
   → sync_http 新增 createSyncHttpClient / obtainSyncHttpClient /
     resetSyncHttpClient，两个工厂都装代理 + 60s 连接超时；顶层同步单例删除
     （它是同步构造的，而代理解析要跑 reg query，这正是它当年没代理的原因）。
     Google auth 三处 + Dropbox/OneDrive/PKCE 15 处调用点改走异步取用。
3. loopback 超时消息带 "authorization"，被 contains('auth') 抢先判成「登录已
   过期」，把用户指向重新登录这条死路。
   → SyncAuthFailureKind 加 browserTimeout；有类型的鉴权失败整体提到字符串猜
     测区之前一次分派。判据是类型，不是字符串。
4. redirect_uri 写 localhost 但服务器只绑 IPv4。localhost 要先过 DNS（Windows
   上先解析 ::1）和代理 bypass 列表，任一不配合授权码就永远回不来。
   → runDesktopOAuthLoopback 加 host 参数，Google 传 127.0.0.1；Dropbox/Entra
     注册的是 localhost 字符串，保持默认值不动。

文案：新增 sync_err_browser_timeout（17 语言）；sync_err_timeout 补「检查网络或
代理设置」；设置项「自定义更新代理」→「自定义网络代理」并写明作用域（持久化 key
冻结不动）。

测试 hibiki/test/sync/oauth_proxy_and_browser_timeout_test.dart（11 例）：代理
配置到达同步层、三种 kind 各说各的话、无误伤原有 401 分类，加 6 条源码守卫；
结构守卫调头禁止代理实现回流 part。6 个守卫全部变异实测见红。
